requirements.txt
matplotlib==3.5.2
matplotlib-inline==0.1.3
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
openpyxl==3.0.10
pandas==1.4.2
plotly==5.6.0
protobuf>=3.20
regex==2022.3.15
scikit-learn==1.1.1
scipy==1.8.0
streamlit>=1.25.0
pyodbc==4.0.34
streamlit-option-menu==0.3.2
pandas-geojson==1.2.0
ydata_profiling==4.7.0
streamlit-pandas-profiling==0.1.3
starrocks
mysql-connector-python
SQLAlchemy